Split widget: signals header
